"Refactor lab problem solution class for readability and maintainability
Detail:
Updated the variable names to more intuitive ones in the '연구소' class to make the code more readable and maintainable. Constants were introduced for better context and understanding. Extraneous 'printMap' and 'printWall' methods which seemed to be for debugging purposes were removed. Comments were corrected for a better explanation. The process of choosing positions for the wall was clarified with a comment indicating that this is a combination algorithm."